Update Makefile.config to autodetect the BASEDIR
Konstantinos Koukos (https://github.com/koukos) sent me a way to do this
automatically, which is better than asking the user to do it. Since I am
not sure if it works on all systems, I am leaving the part about setting
the variable manually in README.md and in the Makefile.