Renamed all .C and .H files to .c.in and .h.in respectively
This is done in order to solve issues caused by case-insensitive file
systems, where file.C and file.c are considered to be the same file. It
also has the advantage that now it's clearer that the .in files are
supposed to be used to generate the actual source files.

The makefiles have been updated, including modernizing the main
Makefile.config to use modern pattern matching instead of suffix rules.