Added ROI markers to the benchmarks
The markers are can be anything, although usually they are function
calls. By default, the ROI markers are empty, allowing for the macros to
be overwritten based on the requirements.

There are two markers:

`SPLASH3_ROI_BEGIN` is placed before all the worker threads are started.
`SPLASH3_ROI_END` is placed after all the worker threads have finished.

Both of these are placed only in the main thread, not on the individual
worker threads.