Fix failing 32bit ARM builds due to array sizes
In the non-continuous ocean benchmark the array sizes are too big for
32bit architectures. In x86 this has been handled by checking for a
commonly found preprocessor macro and switching to smaller sizes. This
patch does the same for ARM as well. This is based on GCC
(crosscompilation), so it might not work on other compilers.

This also fixes the size on 64bit versions to be a power of 2 + 2, as
using other sizes makes no sense in this benchmark.

Fixes #9.