Use `TOOLCHAIN_PREFIX` as xcomp prefix; rm -static.
Based on suggestions by SakalisC, use `TOOLCHAIN_PREFIX` instead of
`CROSS_COMPILE` for the prefix for a cross-compiling toolchain. This is
clearer and more self-explanatory.

The `-static` flag was also removed from `Makefile.config` as it is only
relevant when using gem5 or a similar simulator, not for the benchmarks
in general.